factored out common SR tests to work for both kinds of SR decks
